feat(website)!: remove the readonly and source modes from the homepage
