+++ website/src/components/segmented-control.tsx
@@ -1,0 +1,42 @@
+import type { CSSProperties } from 'react'
+
+interface SegmentedControlProps<T extends string> {
+  options: ReadonlyArray<{ value: T; label: string }>
+  value: T
+  onChange: (value: T) => void
+  ariaLabel?: string
+  /**
+   * Shared radio group name; required when several controls coexist.
+   */
+  name?: string
+}
+
+export function SegmentedControl<T extends string>({
+  options,
+  value,
+  onChange,
+  ariaLabel,
+  name = 'segmented-control',
+}: SegmentedControlProps<T>) {
+  return (
+    <div
+      role="radiogroup"
+      aria-label={ariaLabel}
+      className="segmented"
+      style={{ '--seg-count': options.length } as CSSProperties}
+    >
+      {options.map((option) => (
+        <label key={option.value}>
+          <input
+            type="radio"
+            name={name}
+            value={option.value}
+            checked={option.value === value}
+            onChange={() => onChange(option.value)}
+          />
+          {option.label}
+        </label>
+      ))}
+    </div>
+  )
+}
